Re-add generic_task
