+++ test/automl/test_multiclass.py
@@ -3,8 +3,9 @@ import numpy as np
 import scipy.sparse
 from sklearn.datasets import load_iris, load_wine
 from flaml import AutoML
-from flaml.automl.data import CLASSIFICATION, get_output_from_log
+from flaml.automl.data import get_output_from_log
 from flaml.automl.model import LGBMEstimator, XGBoostSklearnEstimator, SKLearnEstimator
+from flaml.automl.task.task import CLASSIFICATION
 from flaml import tune
 from flaml.automl.training_log import training_log_reader
